Non-algorithmic discrepancies
Found out that 1752 was the switch to gregorian calendar and that before
that the julian calendar followed different rules for leap years (1700
was a leap year in julian for ex). I added a little check for the 128
year error of the julian calendar but did not check how far this holds
true.
The site used to verify results is
https://www.timeanddate.com/calendar/?year=1700&country=1
Final commit for this class